OpenWeather deps.edn rewrite
This PR primarily addresses the fact that the DarkSky API is no longer
operating.  The app now leverages the OpenWeather API.  I also took
this as an opportunity to use deps.edn and shadow-cljs to run the
full-stack app.  The backend and frontend can be run independently for
local dev, while the backend production jar serves up the compiled
frontend for real deployments.

Other changes of note:
- all dependencies have been updated as of 2023-06-22
- styles are now handled within the frontend
- configuration is now handled via the `aero` library from juxt
- weather icons are day / night aware based on sunset time